requirements.txt
Django
djangorestframework
markdown